refactor(ui): split remote host overlay helpers
